Added files from hash_generator directory
hashfinder is a web application that allows users to securely hash files and generate unique digital fingerprints. Users can upload a file to the app, and it will generate a hash using various cryptographic algorithms, including MD5, SHA-1, SHA-256, SHA-384, SHA-512, SHA-3, and BLAKE2b. This can be used to verify the integrity of the file or to compare it to other files to check if they are identical. 

This app is built using the Flask framework for Python, and it uses various Python libraries for cryptographic functions, file handling, and web development. The user interface is simple and intuitive, with a file upload form and a results page that displays the various hashes. 

hashfinder is designed with security in mind, with measures in place to prevent malicious uploads and to protect user data. It also includes a Bandit scan to detect potential security issues in the code. 

Try hashfinder today and see how easy it is to generate hashes for your files!